+++ extensions/pia-controls/src/disconnect.ts
@@ -1,4 +1,4 @@
-import { runPrivateInternetAcessCmd } from "./utils";
+import { disconnectFromPIA } from "./utils";
 export default async () => {
-  await runPrivateInternetAcessCmd("disconnect", "Disconnecting from Private Internet Access...");
+  await disconnectFromPIA();
 };